requirements.txt
flask
tabula-py
pandas
PyMuPDF
pymongo
waitress